requirements.txt
Django==3.2.3
django-bootstrap-form==3.4
djangorestframework==3.12.4
Pillow==5.4.1
pytz==2021.1
reportlab==3.5.13
xhtml2pdf==0.2.5
xlwt==1.3.0
